some stream burst changes idk where i was

+++ include/saint/stream_to_burst_message.h
@@ -702,7 +702,7 @@ namespace gr {
        * class. saint::stream_to_burst_message::make is the public interface for
        * creating new instances.
        */
-      static sptr make(double freq_offset, uint16_t burst_size);
+      static sptr make(double freq_offset, uint32_t burst_size);
       
       virtual void set_freq(double freq_offset) = 0;
     };
